moving flowtypev2 class to v2 model pkg
