+++ src/main/java/com/appdirect/sdk/vendorFields/controller/VendorFieldValidationController.java
@@ -4,7 +4,7 @@ import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
 import static org.springframework.web.bind.annotation.RequestMethod.POST;
 
 import com.appdirect.sdk.vendorFields.converter.FlowTypeV2Converter;
-import com.appdirect.sdk.vendorFields.model.FlowTypeV2;
+import com.appdirect.sdk.vendorFields.model.v2.FlowTypeV2;
 import java.util.List;
 import java.util.Locale;
 import java.util.concurrent.Callable;
